requirements.txt
slack_bolt
aiohttp>=3,<4
python-dotenv
orm
orm[sqlite]